feat: CodeCC 同步代码到release-1.3分支 #3987 fix ci compile error

+++ src/backend/ci/core/plugin/biz-plugin/src/main/kotlin/com/tencent/devops/plugin/dao/PluginGithubCheckDao.kt
@@ -82,7 +82,7 @@ class PluginGithubCheckDao {
                 repositoryConfig.repositoryHashId,
                 repositoryConfig.repositoryName,
                 commitId,
-                checkRunId.toLong(),
+                checkRunId,
                 now,
                 now
             ).execute()
