+++ src/backend/ci/core/plugin/biz-plugin/src/main/kotlin/com/tencent/devops/plugin/service/git/CodeWebhookService.kt
@@ -460,7 +460,7 @@ class CodeWebhookService @Autowired constructor(
                 } else {
                     if (buildNum.toInt() >= record.buildNumber) {
                         scmService.updateGithubCheckRuns(
-                            checkRunId = record.checkRunId.toInt(),
+                            checkRunId = record.checkRunId,
                             projectId = projectId,
                             repositoryConfig = repositoryConfig,
                             name = name,
